More improvements on Parser
